redid the entire file system for clarity
