+++ tuna/cli/core/util.py
@@ -4,6 +4,7 @@ Utility functions for the CLI.
 
 """
 import os
+from tuna.cli.core.constants import BLUE, RESET, RED, WARNING_ICON
 
 
 def clear_terminal():
@@ -22,4 +23,7 @@ def log(icon: str, message: str):
     """
     Logs a message with an icon to the terminal.
     """
-    print(f"[{icon}] {message}")
+    if icon == WARNING_ICON:
+        print(f"{RED}[{icon}]{RESET} {message}")
+    else:
+        print(f"{BLUE}[{icon}]{RESET} {message}")
