+++ tuna/cli/services/fluidstack.py
@@ -14,9 +14,9 @@ import inquirer
 import requests
 from tabulate import tabulate
 from halo import Halo
-from tuna.cli.util import log
-from tuna.cli.constants import CHECK_ICON, CROSS_ICON, INFO_ICON, SSH_KEY
-from tuna.cli.jupyter_fs import connect_lab
+from tuna.cli.core.util import log
+from tuna.cli.core.constants import CHECK_ICON, CROSS_ICON, INFO_ICON, SSH_KEY
+from tuna.cli.services.jupyter_fs import connect_lab
 
 
 def select_gpu(api_key: str) -> dict:
